deployment: time out stuck deployment after 1h
Currently, if a deployment gets stuck (ex. the components-api is
restarted, and the deployment task disappears) the deployment will be
kept in 'running' forever.

This adds a cleanup when doing any action that will check if any of
the deployments are stuck, and time them out. Similar to how we
cleanup old deployments.

This adds a new deployment state too, `timed_out`.

Fixed a couple issues also found during the addition of the tests.

Signed-off-by: David Caro <dcaro@wikimedia.org>