runtime: refactor some of the code to a runtime module
This does not change any functionality, just moves around some code
that was all in the deploy task, to a new runtime module that
contains the code to interact with the runtime (toolforge) and a
helper mock runtime for tests.

I had to change a little bit the settings functions to help mocking
for the tests.

With this now we can also simplify a lot of the tests, but that will
be done in a different patch.

Signed-off-by: David Caro <dcaro@wikimedia.org>