openapi spec: ToolConfig-Output config_version is nullable
Per https://swagger.io/docs/specification/v3_0/data-models/data-types/#null
this is the correct way to show the type is nullable.
I imagine this is currently auto generated from code, and needs
to be tweaked elsewhere? But making this commit as a placeholder.

I had to make this alteration to the spec so as to generate a
valid client for toolforge-gen-cli that didnt expect the
config_version to be set in the response, as it doesn't seem
to be there anymore.