Add validated type for git urls
Currently `repository` accepts any string, which we pass through to
`git` and the builds API.

Git supports multiple (transport) protocols including SSH, here we are
only interested in those that do not require authentication.

Validate the URL to be one of `http://`, `https://`, `git://`.

This is similar to the validation performed on `source_url`.

Update the tests to pass valid git urls.