Ensure reuse_from components are re-run
Bug: T403285

When a `component` specifies `reuse_from`, the `build` will be skipped.

When calculating the jobs that require re-running i.e. restarting,
we don't include the components which use a reference as their build
status is `skipped`.

In the case that the `source` is a reference, check the build status of
the "parent" component.