Expand reuse_from validation
Bug: T403287

Currently we assert that a referenced component exists, however the
referenced component can be a reference to another component.

We do not recursively resolve the references and to do that today would
be somewhat complicated due to how the build info is looked up.

Expand the validation to assert that the "parent" component i.e. that
specified in `SourceReference` has a source type of `SourceBuild`.

Also add a test case for this.