Add test for component restart
This is a bit convoluted as the tests currently return the builds
response, which gets loaded into an empty jobs response.

As there is no matching component name in the empty jobs response, the
delete/create (restart) logic is skipped.

Add in a mocked response which triggers the logic, then check the
relevant calls where made to the API.

Really this logic should be refactored to make it easier to test, rather
than embedding the conditional blocks inside e.g. _do_run