Remove `job` from `JobsJobResponse`
This is not returned from jobs-api for either the `PATCH` or `DELETE`
method, so in production it is never set.

We currently set it in tests, though the tests are hitting a code path
that doesn't actually get used in the service.

Drop the field and add in missing response messages, which is what
actually gets returned to the user in production today.