Use the image name as provided by builds-api
When performing a deployment, we specify the `image_name` that a build
should be published at.

Previously we would use this same name to infer the destination image
path.

While this works it implicitly ties builds-api and components-api
together, resulting in the destination namespace and url format being
spread throughtout code.

This change replaces the abstract `get_build_state` with `get_build_info`,
extending the internal `DeploymentBuildInfo` to also hold the
`build_image`.

The `destination_image` key is already set by builds-api, making this a
noop today.